comportamento das fachadas

+++ ReservaOps/src/br/ufal/ic/ppgi/smartagenda/reservaops/spec/req/IReqSistema.java
@@ -4,8 +4,8 @@ import br.ufal.ic.ppgi.smartagenda.reservaops.spec.dt.Usuario;
 
 public interface IReqSistema {
 	
-	public boolean verificarSeUsuarioExiste(Long idUsuario);
+	public boolean verificarSeUsuarioExiste(int idUsuario);
 	
-	public Usuario recuperarUsuario(Long idUsuario);
+	public Usuario recuperarUsuario(int idUsuario);
 
 }
